v2.6.1: numero_amostra INTEGER + cleanup_historico + CHANGELOG
Database:
- Cria migration v4: numero_amostra NUMERIC -> INTEGER
- Atualiza schema.sql e DIAGRAM.MD

Backend (amostra.py):
- Troca safe_float por safe_int (rejeita decimais)
- Mensagens de erro atualizadas para 'inteiro positivo'
- Lote usa safe_int em vez de int() manual

Frontend (amostra.html):
- step='any' -> step='1' e min='0' -> min='1'

Testes:
- cleanup_historico fixture removendo historico_log apos testes
- Adicionado cleanup_historico em 6 testes create_*
- test_create_amostra: valor 1.5 -> 2 (float nao aceito)

CHANGELOG.md e README atualizados

+++ database/schema.sql
@@ -75,7 +75,7 @@ CREATE TABLE historico_log (
 -- Tabela: amostra
 CREATE TABLE amostra (
     id SERIAL PRIMARY KEY,
-    numero_amostra NUMERIC NOT NULL,
+    numero_amostra INTEGER NOT NULL,
     lote INTEGER NOT NULL,
     user_id UUID REFERENCES auth.users(id),
     created_at TIMESTAMP DEFAULT NOW(),
